ImageProcessingService/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
pillow==11.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytest==8.4.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
